+++ sql/schemas/hapi-pop-tables.sql
@@ -5,12 +5,14 @@
 --
 -- These depend on the tables in hapi-core-tables.sql
 --
--- For ANSI SQL, replace "SERIAL PRIMARY KEY" with "GENERATED ALWAYS
--- AS IDENTITY" (works with PostgreSQL, but not SQLite or MySQL).
+-- For auto incrementing primary keys, use:
+-- PostgreSQL: SERIAL PRIMARY KEY
+-- sqlite: INTEGER PRIMARY KEY
+-- ANSI SQL: GENERATED ALWAYS AS IDENTITY
 -- ---------------------------------------------------------------------
 
 CREATE TABLE Population (
-       id SERIAL PRIMARY KEY,
+       id INTEGER PRIMARY KEY,
        resource_ref INT NOT NULL,
        admin2_ref INT NOT NULL,
        gender_code CHAR(1),